Fix a bug in HttpdChecker
When it was found at a URL, other URLs weren't checked. The issue is
sometimes it's found without the version and another URL contains the
technology and the version, but it wasn't checked.
Now it's checked on all URLs except when the version is found, in this
case the finding is returned directly.